fack: initial version
This is the initial code commit for the fack tool.

It checks a file's last-modified date against a configurable timespan or
displays the last timespan as human-readable string. The tool also
supports a silent mode (--non-interactive) in which it returns 0 on a current
file and -1 on an outdatet file without printing anything on the shell (except of
errors in the command line argument parsing), -2 in case of an error

Signed-off-by: Johannes Zink <opensource@johannes-zink.de>